refactor program and instance content types and move all to types folder
