+++ src/messages/types/base.ts
@@ -4,8 +4,8 @@
  *
  * Warning: Avax, CSDK, NEO are currently not supported by the TS sdk.
  */
-import { InstanceContent } from "../instance/types";
-import { ProgramContent } from "../program/programModel";
+import { InstanceContent } from "./instance";
+import { ProgramContent } from "./program";
 
 export enum Chain {
     DOT = "DOT",
