+++ src/messages/types/index.ts
@@ -1,3 +1,5 @@
 export * from "./base";
 export * from "./volumes";
 export * from "./execution";
+export * from "./program";
+export * from "./instance";
