+++ src/messages/types/program.ts
@@ -1,5 +1,4 @@
-import { BaseContent, MachineVolume } from "../types";
-import { FunctionEnvironment, MachineResources, Chain, PaymentType } from "../types";
+import { BaseExecutableContent } from "./execution";
 
 /**
  * Type of Encoding
@@ -18,6 +17,14 @@ export enum MachineType {
 }
 
 /**
+ * Type of interface
+ */
+export enum InterfaceType {
+    asgi = "asgi",
+    binary = "binary",
+}
+
+/**
  * Code to execute
  */
 export type CodeContent = {
@@ -25,8 +32,8 @@ export type CodeContent = {
     entrypoint: string;
     ref: string;
     use_latest: boolean;
-    interface: string;
-    args: string[];
+    interface?: InterfaceType;
+    args?: string[];
 };
 
 /**
@@ -65,28 +72,11 @@ export type FunctionRuntime = {
     comment: string;
 };
 
-/**
- * Payment solution
- */
-export type Payment = {
-    chain: Chain;
-    receiver?: string;
-    type: PaymentType;
-};
-
-export type ProgramContent = BaseContent & {
-    type: MachineType;
-    allow_amend: boolean;
+export type ProgramContent = BaseExecutableContent & {
+    type: MachineType.vm_function;
     code: CodeContent;
-    variables?: { [key: string]: string };
+    runtime: FunctionRuntime;
     data?: DataContent;
     export?: Export;
     on: FunctionTriggers;
-    metadata?: Record<string, any>;
-    environment: FunctionEnvironment;
-    resources: MachineResources;
-    runtime: FunctionRuntime;
-    volumes: MachineVolume[];
-    payment?: Payment;
-    replaces?: string;
 };
